Rename modelView to appView
