+++ examples/hosts/app-integration/schema-upgrade/src/view/appView.tsx
