+++ examples/hosts/app-integration/schema-upgrade/src/view/index.ts
@@ -4,4 +4,4 @@
  */
 
 export * from "./debugView";
-export * from "./modelView";
+export * from "./appView";
